+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## Cassava Megaparsec 1.0.0
+
+* Works with Megaparsec 6.
+
+* Instead of `Cec` we have `ConversionError` as custom component of error
+  messages.
+
 ## Cassava Megaparsec 0.1.0
 
 * Initial release.